requirements.txt
grpcio==1.8.3
grpcio-tools==1.8.3
protobuf==3.5.1
PyMySQL==0.8.0
redis==2.10.6
schedule==0.5.0
six==1.11.0
